+++ mg/framework/annotation/Get.java
@@ -1,0 +1,6 @@
+package mg.framework.annotation;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Get {}
