+++ mg/framework/annotation/RequestParam.java
@@ -1,0 +1,9 @@
+package mg.framework.annotation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RequestParam {
+    String value();
+}
