+++ mg/framework/annotation/RestAPI.java
@@ -1,0 +1,7 @@
+package mg.framework.annotation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RestAPI {}
