+++ mg/framework/annotation/Url.java
@@ -1,0 +1,8 @@
+package mg.framework.annotation;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Url {
+    String value();
+}
