+++ mg/framework/annotation/validation/Required.java
@@ -1,0 +1,7 @@
+package mg.framework.annotation.validation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Required {}
