+++ mg/framework/exception/PackageException.java
@@ -1,0 +1,7 @@
+package mg.framework.exception;
+
+public class PackageException extends Exception{
+    public PackageException(String packageName) {
+        super("The package " + packageName + " is empty or doesn't exist!");
+    }
+}
