+++ mg/framework/exception/UrlNotFoundException.java
@@ -1,0 +1,10 @@
+package mg.framework.exception;
+
+public class UrlNotFoundException extends Exception{
+    public UrlNotFoundException () {
+        super("404 not found!");
+    }
+    public UrlNotFoundException (String message) {
+        super(message);
+    }
+}
