--- mg/itu/framework/sprint/annotation/Controller.java
@@ -1,7 +1,0 @@
-package mg.itu.framework.sprint.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Controller {}
