--- mg/itu/framework/sprint/annotation/Post.java
@@ -1,6 +1,0 @@
-package mg.itu.framework.sprint.annotation;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Post {}
