--- mg/itu/framework/sprint/annotation/RequestParam.java
@@ -1,9 +1,0 @@
-package mg.itu.framework.sprint.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequestParam {
-    String value();
-}
