--- mg/itu/framework/sprint/annotation/Url.java
@@ -1,8 +1,0 @@
-package mg.itu.framework.sprint.annotation;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Url {
-    String value();
-}
