--- mg/itu/framework/sprint/exception/CastException.java
@@ -1,10 +1,0 @@
-package mg.itu.framework.sprint.exception;
-
-public class CastException extends Exception{
-    public CastException(String message) {
-        super(message);
-    }
-    public CastException() {
-        super("Can't cast object!");
-    }
-}
