--- mg/itu/framework/sprint/servlet/ModelView.java
@@ -1,33 +1,0 @@
-package mg.itu.framework.sprint.servlet;
-
-import java.util.HashMap;
-
-public class ModelView {
-    String url;
-    HashMap<String,Object> data = new HashMap<>();
-    
-    public ModelView(){}
-    public ModelView(String url, HashMap<String, Object> data) {
-        this.url = url;
-        this.data = data;
-    }
-    public ModelView(String url) {
-        this.url = url;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-    public void setUrl(String url) {
-        this.url = url;
-    }
-    public HashMap<String, Object> getData() {
-        return data;
-    }
-    public void setData(HashMap<String, Object> data) {
-        this.data = data;
-    }
-    public void addData(String varName, Object varValue) {
-        this.data.put(varName, varValue);
-    }
-}
